infrastructure/ai-knowledge-agent/requirements.txt
# =================================================================
# AI Knowledge Agent - Python Dependencies
# =================================================================
# CPU-optimized for resource efficiency
# =================================================================

# PyTorch - CPU-only version (saves ~2GB vs CUDA version)
--index-url https://download.pytorch.org/whl/cpu
torch==2.1.0

# Sentence Transformers - Pre-trained embedding models
sentence-transformers==2.2.2

# PostgreSQL with pgvector support
psycopg2-binary==2.9.9
pgvector==0.2.3

# Database connection pooling
sqlalchemy==2.0.23

# HTTP server for health checks
flask==3.0.0

# Logging
loguru==0.7.2

# Environment variables
python-dotenv==1.0.0

# Numpy (included but specified for version lock)
numpy==1.24.3
